+++ app/src/androidTest/java/com/vce/tsunagi/BackfillTest.kt
@@ -49,6 +49,7 @@ class BackfillTest {
         }
 
         override fun recordSweep(at: Long) = Unit
+        override fun clearExclusionWindow() = Unit
     }
 
     @Before